+++ CHANGES.md
@@ -1,6 +1,7 @@
 # Changes in HTMLDOC v1.9.1
 
 - Fixed monospace font size issue (Issue #309)
+- Added support for reproducible builds (Issue #310)
 - Added limited support for the HTML 4.0 SPAN element (Issue #311)
 - Added (extremely limited) UTF-8 support for input files (Issue #314)
 - Now indent PRE text, by popular request.